Creating branch which only has streamlit app
