+++ environment_windows.yml
@@ -1,114 +1,17 @@
+# Before creating the conda env, install ImageMagick from here: https://imagemagick.org/script/download.php
 name: rillgen2d
 channels:
   - conda-forge
   - defaults
   - anaconda
 dependencies:
-  - affine
-  - attrs
-  - boost-cpp
-  - branca
-  - brotlipy
-  - bzip2
-  - ca-certificates
-  - cairo
-  - certifi
-  - cffi
-  - cfitsio
-  - chardet
-  - click
-  - click-plugins
-  - cligj
-  - cryptography
-  - curl
-  - cycler
-  - expat
   - folium
-  - fontconfig
-  - freetype
-  - freexl
   - gdal
-  - geos
-  - geotiff
-  - gettext
-  - hdf4
-  - hdf5
-  - icu
-  - idna
-  - intel-openmp
-  - jinja2
-  - jpeg
-  - kealib
-  - kiwisolver
-  - krb5
-  - libblas
-  - libcblas
-  - libclang
-  - libcurl
-  - libffi
-  - libgdal
-  - libglib
-  - libiconv
-  - libkml
-  - liblapack
-  - libnetcdf
-  - libpng
-  - libpq
-  - libspatialite
-  - libssh2
-  - libtiff
-  - libwebp-base
-  - libxml2
-  - lz4-c
   - m2w64-gcc
-  - markupsafe
-  - matplotlib
-  - matplotlib-base
-  - mkl
-  - numpy
-  - olefile
-  - openjpeg
-  - openssl
-  - pcre
+  - m2-gawk
+  - m2-coreutils
   - pillow
+  - streamlit>=1.18
   - pip
-  - pixman
-  - poppler
-  - poppler-data
-  - postgresql
-  - proj
-  - pycparser
-  - pyopenssl
-  - pyparsing
-  - pyqt
-  - pyqt-impl
-  - pyqt5-sip
-  - pyqtchart
-  - pyqtwebengine
-  - pysocks
-  - python
-  - python-dateutil
-  - python_abi
-  - qt
-  - rasterio
-  - requests
-  - setuptools
-  - sip
-  - six
-  - snuggs
-  - sqlite
-  - tiledb
-  - tk
-  - tornado
-  - urllib3
-  - vc
-  - vs2015_runtime
-  - wheel
-  - win_inet_pton
-  - wincertstore
-  - xerces-c
-  - xz
-  - zlib
-  - zstd
   - pip:
     - wand
